+++ extensions/project-creator/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.0.2
+
+- fix: rax-spa project targets null problem. [#861](https://github.com/appworks-lab/pack/issues/861)
+- fix .vscodeignore configuration [#576](https://github.com/microsoft/vscode-vsce/issues/576)
+
+## 1.0.1
+
+- chore: update webview icon
+
 ## 1.0.0
 
 - chore: release 1.0.0, Rename to Appworks