+++ extensions/style-helper/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.0.1
+- feat: support *.less style.
+- feat: upgrade [css](https://www.npmjs.com/package/css) fix `@media` parse failed.
+
 ## 1.0.0
 
 Release 1.0.0